Render as a plain graph

This renderer uses viz.js to render the process map using the DOT layout.

renderer_graphviz(
  svg_fit = TRUE,
  svg_contain = FALSE,
  svg_resize_fit = TRUE,
  zoom_controls = TRUE,
  zoom_initial = NULL
)

Arguments

svg_fit

Whether to scale the process map to fully fit in its container. If set to `TRUE` the process map will be scaled to be fully visible and may appear very small.

svg_contain

Whether to scale the process map to use all available space (contain) from its container. If set to `FALSE`, if `svg_fit` is set this takes precedence.

svg_resize_fit

Whether to (re)-fit the process map to its container upon resize.

zoom_controls

Whether to show zoom controls.

zoom_initial

The initial zoom level to use.

Value

A rendering function to be used with animate_process

See also

Examples

data(example_log)

# Animate the process with the default GraphViz DOT renderer
animate_process(example_log, renderer = renderer_graphviz())